chore: chose parser feat: add health check to compose files so db-api starts AFTER sql container is ready
