+++ server/docker-compose.yml
@@ -28,13 +28,19 @@ services:
       context: ./bedfellow-db-api/
       dockerfile: Dockerfile
     depends_on:
-      - bedfellow-db
+      bedfellow-db:
+        condition: service_healthy
     ports:
       - 127.0.0.1:8000:8000
     expose:
       - 8000
   bedfellow-db:
     image: mysql:latest
+    healthcheck:
+      test: ["CMD", 'mysqladmin', 'ping', '-h', 'localhost', '-u', 'root', '-p$$MYSQL_ROOT_PASSWORD' ]
+      interval: 2s
+      timeout: 60s
+      retries: 20
     restart: always
     environment:
       MYSQL_DATABASE: bedfellow
