+++ src/types/index.ts
@@ -1,5 +1,9 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 
+export * as BedfellowTypes from './bedfellow-api';
+export * as WhoSampledTypes from './whosampled';
+
+// TODO re-arrange these later get the rn nac stack stuff out this should just be datatypes
 export type RootStackParamList = {
   Login: undefined;
   Details: undefined;
@@ -14,48 +18,3 @@ export type DetailsScreenProps = NativeStackScreenProps<
   RootStackParamList,
   'Details'
 >;
-
-export interface WhoSampledData {
-  track_name: string;
-  artist: string;
-  year: number;
-  image: string;
-}
-
-export interface WhoSampledSearchData {
-  id: number;
-  url: string;
-  artist_name: string;
-  track_name: string;
-  image_url: string;
-  counts: string;
-}
-
-export interface WhoSampledResponse {
-  samples: Array<WhoSampledData>;
-  sampled_by: Array<WhoSampledData>;
-  covers: Array<WhoSampledData>;
-}
-
-export interface BedfellowSample {
-  artist: String;
-  id: Number;
-  image: Blob;
-  track: String;
-  year: 1995;
-}
-
-export interface BedfellowTrackSamples {
-  artist: String;
-  samples: Array<BedfellowSample>;
-  status: String;
-  track: String;
-}
-
-export interface WhoSampledSearchResponse {
-  tracks: Array<WhoSampledSearchData>;
-}
-
-export interface WhoSampledErrorResponse {
-  error: string;
-}
